Cast (size_t)(char) for isprint(), isspace(), isdigit(), toupper(), etc. who may be macros and use it as array subscript (due to sys/ctype_inline.h)

+++ drivers/apcsmart-old.c
@@ -536,7 +536,7 @@ static void protocol_verify(unsigned char cmd)
 	if (found)
 		return;
 
-	if (isprint(cmd))
+	if (isprint((size_t)cmd))
 		upsdebugx(1, "protocol_verify: 0x%02x [%c] unrecognized",
 			cmd, cmd);
 	else
